Changed number of Max_gateways from 7 to 5 and increased max_stargates from 3 - 4 - Going for more lategame Air strategy.
Added near_by_enemy_threats for defending. Want to only grab enemy threats within 20 units of a nexus.

Changed placement for building_assimilators to higher on the ordering.

Added in a quick minerals check before attemping to build any buildings or units.

Added in targets kwarg to find_target to grab a random target from a subgroup of units instead of all units.

changed defend to check near_by_enemy_threats instead of just enemy_threats, and if defending is called for we pass near_by_enemy_threats to find_target

changed logic for regroup. still not being used, but once I'm more comfortable with a location to regroup to before attacking I will add it in.